+++ src/BACK/barbers/barbers.router.ts
@@ -1,15 +1,15 @@
 import express from "express";
 const router = express.Router();
 
-import * as controller from "./barberos.controller";
+import * as controller from "./barbers.controller";
 
-router.get("/createBarberos", controller.create);
+router.get("/createBarbers", controller.create);
 router.post("/", controller.store);
 
 router.get("/", controller.index);
 router.get("/:cuil", controller.show);
 
-router.get("/:cuil/updateBarbero", controller.edit);
+router.get("/:cuil/updateBarber", controller.edit);
 router.put("/:cuil", controller.update);
 
 router.delete("/:cuil", controller.destroy);
