+++ src/BACK/categories/categories.router.ts
@@ -1,15 +1,15 @@
 const express = require("express");
 const router = express.Router();
 
-const controller = require("../controllers/categorias.controller");
+const controller = require("../controllers/categories.controller");
 
-router.get("/createCategorias", controller.create);
+router.get("/createCategories", controller.create);
 router.post("/", controller.store);
 
 router.get("/", controller.index);
 router.get("/:codCategoria", controller.show);
 
-router.get("/:codCategoria/modificarCategorias", controller.edit);
+router.get("/:codCategoria/updateCategories", controller.edit);
 router.put("/:codCategoria", controller.update);
 
 router.delete("/:codCategoria", controller.destroy);
