+++ src/BACK/typeOfHaircut/typeOfHaircut.router.ts
@@ -1,18 +1,18 @@
 import express from "express";
 const router = express.Router(); // enrutador para la seccion tipoCortes
 
-import * as controller from "./tipoCorte.controller"; // importar controlador de tipoCorte
+import * as controller from "./typeOfHaircut.controller"; // importar controlador de tipoCorte
 //get: leer datos
 //post: enviar y guardar datos
 //put: actualizar datos
 
-router.get("/createTipoCorte", controller.create);
+router.get("/createTypeOfHaircut", controller.create);
 router.post("/", controller.store);
 
 router.get("/", controller.index);
 router.get("/:codCorte", controller.show);
 
-router.get("/:codCorte/modificarTipoCorte", controller.edit);
+router.get("/:codCorte/modificarTypeOfHaircut", controller.edit);
 router.put("/:codCorte", controller.update);
 
 router.delete("/:codCorte", controller.destroy);
