+++ src/FRONT/views/components/categories/createCategories.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import styles from "./categorias.module.css";
+import styles from "./categories.module.css";
 
 const createCategorias: React.FC = () => {
   const navigate = useNavigate();
