+++ src/FRONT/views/components/typeOfHaircut/typeOfHaircut.module.css
